Was made complete logic with categories and tested all classes.

+++ src/main/java/com/example/repositories/CategoryRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    List<Category> findByParentCategoryId(int parentCategoryId);
+    List<Category> findByParentCategoryId(long parentCategoryId);
 
     List<Category> findByParentCategoryIdNotNull();
 }
